Bit of refactoring, more tests.

+++ txcache/constants.go
@@ -10,4 +10,4 @@ const numEvictedTxsToDisplay = 3
 
 const excellentGasPriceFactor = 5
 
-const numberOfScoreChunks = uint32(100)
+const numberOfScoreChunks = 100
